Stop depending on core-system-modules
The core-system-modules provides access to all the implementation
details of the core libraries which is unnecessary for this. This
change switches to use art-module-public-api-stubs-system-modules
which is the part of the public API provided by libcore and so is much
more limited and is also very stable.

As art-module-public-api-stubs-system-modules does not contain
conscrypt or bouncycastle it was necessary to add those as extra
dependencies.

Test: m checkbuild
Bug: 141613583
Change-Id: Icf2b7d18521b1ab601727c5ddc7d511f4066717d