Wycheproof: Upstream changes for EcdsaTest
Upstream changes from Wycheproof project with support for
AndroidKeystore provider and keys. The message from Upstream CL as
follows -
"Changing EcdsaTest, so that it works with deterministic ECDSA.
Adding a test that checks the behaviour of ECDSA when null is passed as
SecureRandom argument. If the implementation is randomized otherwise
then we still expect randomized signatures. If the implementation is
deterministic, we just expect this to work and not throw a NullPointerException.

Testing: I'm testing this against openjdk and BouncyCastle. BouncyCastle has
a deterministic implementation of ECDSA, which is called ECDDSA. Providers that
don't know ECDDSA (which is every provider except BouncyCastle just skip the test)."

Bug: 258011138
Test: atest CtsKeystoreWycheproofTestCases:com.google.security.wycheproof.EcdsaTest
Change-Id: I1c03b8e14b9a7cd8055ef470c9558bbc30108958